Do not use separate binaries
